+++ Changes.md
@@ -1,3 +1,17 @@
+0.57.x.x (relative to 0.57.5.0)
+========
+
+Fixes
+-----
+
+- SceneAlgo : Removed cancellers from contexts referenced by history objects.
+- Context : Fixed Python `Canceller` lifetime management bug.
+
+API
+---
+
+- Context : Added copy constructor that allows an existing canceller to be omitted.
+
 0.57.5.0 (relative to 0.57.4.1)
 ========